requirements.txt
scikit-learn
matplotlib
pandas 
numpy
pyyaml
dvc[s3]